requirements.txt
PyPDF2==3.0.1
pandas>=2.0.0
scikit-learn>=1.0.0
nltk>=3.0.0
spacy>=3.0.0
matplotlib
seaborn
PyYAML
torch
transformers